Added functionality of the game board.
Added a class to represent the core functionality of the connect 4 game
board. Players can choose a column and drop a token. If a column is chosen
that is not allowed, an exception will be thrown. It is recommended at
this point that a player making such a move will lose automatically. A
main method has also been added, which displays the core functionality of
the game, by displaying a prescripted game in which one of the players
wins.