Added an interface for Connect4Board
Added an interface for Connect4Board, specifying methods that do not
modify the internals of the Connect4Board object and only fetch
information about the Connect 4 board. This interface replaced the
parameter of Agent.perceive(board). This allows implementations of Agent
to get row and column information about a connect 4 board directly,
without needing to call getCopyOfBoard(), which was creating a new
PlayerColor[][] object on every turn. This also allowed the
PrimitivePreventionAgent to be optimized, as that Agent implementation was
calling Connect4Board.winner(board) on every decision. Now the agent
performs its own determinations for whether or not a player will win given
possible future board states. Some additional changes were made to Main to
accomodate this change.