Moved methods to Agent from various sources.
While adding threat detection to the PerfectAgent, realized that it needed
functionality that was already implemented by the
PrimitivePreventionAgent. Those functionalities were moved to Agent.

The IConnect4Board functionality that determines how many emtpy rows are
in a column was moved to Agent. The initial reasoning for adding this
functionality ended up applying more to the series of Agent
implementations.

In the future, methods will be more rigorously analyzed for their proper
place in the codebase.